RequireJS Module loading with demo modules to load
require.html is the html page with module loading. The list of modules to load is defined in modules/moduleList.js and all other modules are also placed in the modules folder. The modules follow a pattern that needs to be held to while creating new modules. Documentation for it will follow.